+++ project/src/store/actions/authorizationActions.ts
@@ -1,11 +1,72 @@
 import { UserActionType } from '../../const';
 import UserState from '../../types/userState';
+import { AuthorizationStatus, EndPoint } from '../../const';
+import { dropToken, saveToken } from '../../services/token';
+import { AuthData } from '../../types/auth-data';
+import LoginResponse from '../../types/loginResponse';
+import { ThunkAction } from 'redux-thunk'; //ThunkDispatch
+import { State } from '../../types/state';
+import { AxiosInstance } from 'axios';
 
-export const requireAuthorization = (authData: UserState) => ({
-  type: UserActionType.Login,
-  payload: authData,
-} as const);
+export type UserActions =
+  | ReturnType<typeof requireAuthorization>
+  | ReturnType<typeof requireLogout>;
 
-export const requireLogout = () => ({
-  type: UserActionType.Logout,
-} as const);
+export const requireAuthorization = (authData: UserState) =>
+  ({
+    type: UserActionType.Login,
+    payload: authData,
+  } as const);
+
+export const requireLogout = () =>
+  ({
+    type: UserActionType.Logout,
+  } as const);
+
+/* Async Actions */
+type ThunkActionResult<R = Promise<void>> = ThunkAction<
+  R,
+  State,
+  AxiosInstance,
+  UserActions
+>;
+
+export const checkAuthAction =
+  (): ThunkActionResult => async (dispatch, _getState, api) => {
+    await api.get<LoginResponse>(EndPoint.Login).then((response) => {
+      response.data && dispatch(requireAuthorization(adaptLoginResponse(response.data)));
+    });
+  };
+
+export const loginAction =
+  ({ login: email, password }: AuthData): ThunkActionResult =>
+    async (dispatch, _getState, api) => {
+      const { data } = await api.post<LoginResponse>(EndPoint.Login, {
+        email,
+        password,
+      });
+      saveToken(data.token);
+
+      dispatch(requireAuthorization(adaptLoginResponse(data)));
+    };
+
+export const logoutAction =
+  (): ThunkActionResult => async (dispatch, _getState, api) => {
+    api.delete(EndPoint.Logout);
+    dropToken();
+    dispatch(requireLogout());
+  };
+
+/* Utils */
+
+function adaptLoginResponse(data: LoginResponse): UserState {
+  const adaptedData: any = {
+    ...data,
+    avatarURL: data['avatar_url'],
+    status: AuthorizationStatus.Auth,
+  };
+  delete adaptedData.token;
+  delete adaptedData['avatar_url'];
+
+  return adaptedData;
+}
