+++ project/src/store/reducers/userReducer.ts
@@ -1,12 +1,12 @@
 import { AuthorizationStatus, UserActionType } from '../../const';
-import { Actions } from '../../types/action';
 import type UserState from '../../types/userState';
+import { UserActions } from '../actions/authorizationActions';
 
 const InitialState: UserState = {
   status: AuthorizationStatus.Unknown,
 };
 
-function userReducer(state = InitialState, action: Actions): UserState {
+function userReducer(state = InitialState, action: UserActions): UserState {
   switch (action.type) {
     case UserActionType.Login:
       return action.payload;
